Rename file to index.tsx
